Moved form code into separate components to distinguish difference between court and non-court related disputes
