+++ client/src/views/CaseForm.js
@@ -57,84 +57,7 @@ if (process.env.NODE_ENV === 'production') {
 
 
 const CaseForm = (props) =>{
-  const classes = useStyles();
   
-      const [form, setValues] = useState({
-        "parties_involved":"",
-        "description":"",
-        "dispute_category":""
-      });
-      
-      
-      const handleChange = name => event => {
-        setValues({ ...form, [name]: event.target.value });
-      };
-      
-      const onSubmitHandler = async e =>{
-        e.preventDefault();
-        console.log(form,"form")
-        await axios
-        .post(`${process.env.REACT_APP_API_URL}/cases`,form,{
-          headers: {
-            "Authorization": localStorage.getItem('token'),
-          },})
-        .then(res => {
-          console.log("add new case: ", res.data)
-        })
-        .catch(err => {
-          console.log("add case err", err.response)
-        })
-        props.history.push("/home");
-      }
-
-
-    return (
-
-      <div style={{maxWidth:"1100px",margin:"0 auto",padding:"100px"}}>
-
-          <h1 style={{textAlign:"center"}}>Case form</h1>
-        
-          <form className={classes.container} noValidate autoComplete="off" onSubmit={onSubmitHandler}>
-
-              <Select
-                value={form.dispute_category}
-                onChange={handleChange("dispute_category")}
-              >
-                <MenuItem value="Landlord/Tenant">Landlord/Tenant</MenuItem>
-                <MenuItem value="Eldercare">Eldercare</MenuItem>
-                <MenuItem value="Commercial">Commercial</MenuItem>
-                <MenuItem value="Domestic">Domestic</MenuItem>
-                <MenuItem value="Workplace">Workplace</MenuItem>
-                <MenuItem value="Penal">Penal</MenuItem>
-            </Select>
-
-              <TextField
-              label="Third Party Email"
-              type="email"
-              name="parties_involved"
-              autoComplete="email"
-              margin="normal"
-              variant="outlined"
-              onChange={handleChange("parties_involved")}
-              value={form.parties_involved}  
-            />
-
-              <TextField
-              name="description"
-              label="Description of Conflict"
-              multiline
-              rowsMax="8"
-              value={form.description}
-              onChange={handleChange("description")}
-              margin="normal"
-              variant="outlined"
-              />
-
-              <Button className={classes.button} onClick={onSubmitHandler}>submit</Button>
-
-          </form>
-      </div>
-    )
   };
   
   /**
